[M1_TR-212] Refactored structure
