+++ web/src/app/job/context.ts
@@ -1,0 +1,4 @@
+import { JobTable } from '@/utils/types/job';
+import { createContext } from 'react';
+
+export const JobListContext = createContext<JobTable | undefined>(undefined);
