+++ web/src/components/molecules/EstimationStatusFilter/hooks.ts
@@ -1,0 +1,15 @@
+import { SelectChangeEvent } from '@mui/material';
+import { useState } from 'react';
+
+export const useHooks = () => {
+	const [status, setStatus] = useState('');
+
+	const handleChange = (e: SelectChangeEvent) => {
+		setStatus(e.target.value);
+	};
+
+	return {
+		status,
+		handleChange
+	};
+};
