+++ web/src/components/organisms/SearchFilterHeader/hooks.ts
@@ -1,0 +1,14 @@
+import { useState } from 'react';
+
+export const useHooks = () => {
+    const [isExpanded, setIsExpanded] = useState(false);
+
+	const toggleFilters = () => {
+		setIsExpanded(!isExpanded);
+	};
+
+    return {
+        isExpanded,
+        toggleFilters
+    };
+};
