+++ web/src/utils/constants/estimationStatus.ts
@@ -1,0 +1,7 @@
+export const EstimationStatus = [
+	{ id: 1, status: 'Not yet Created' },
+	{ id: 2, status: 'Making' },
+	{ id: 3, status: 'Approved' },
+	{ id: 4, status: 'Sent to Customer' },
+	{ id: 5, status: 'Closed' }
+];
