+++ web/src/utils/constants/statusChipColor.ts
@@ -1,0 +1,11 @@
+export interface ChipProps {
+	label: string;
+}
+
+export const ChipColors: Record<string, string> = {
+	'Not yet Created': '#FFB4AF',
+	Making: '#FDFF8F',
+	Approved: '#8AFFB2',
+	'Sent to Customer': '#84C1FF',
+	Closed: '#65707b33'
+};
