+++ web/src/utils/constants/tags.ts
@@ -1,0 +1,5 @@
+export const Tags = [
+	{ id: 1, value: 'TAG_A', name: 'Tag A' },
+	{ id: 2, value: 'TAG_B', name: 'Tag B' },
+	{ id: 3, value: 'TAG_C', name: 'Tag C' }
+];
